requirements.txt
anyio==3.6.1
asyncpg==0.26.0
click==8.1.3
fastapi==0.85.0
greenlet==1.1.3
gunicorn==20.1.0
h11==0.14.0
idna==3.4
MarkupSafe==2.1.1
Pillow==9.2.0
psycopg2-binary==2.9.4
pydantic==1.10.2
PyMySQL==1.0.2
python-dotenv==0.21.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.18.3
Werkzeug==2.2.2
